Refactor and move definitions and value objects to Models namespace
