+++ src/XmlCancelacionHelper.php
@@ -5,15 +5,15 @@ declare(strict_types=1);
 namespace PhpCfdi\XmlCancelacion;
 
 use DateTimeImmutable;
-use PhpCfdi\XmlCancelacion\Capsules\CancelDocument;
-use PhpCfdi\XmlCancelacion\Capsules\CancelDocuments;
+use PhpCfdi\XmlCancelacion\Models\CancelDocument;
+use PhpCfdi\XmlCancelacion\Models\CancelDocuments;
 use PhpCfdi\XmlCancelacion\Capsules\Cancellation;
 use PhpCfdi\XmlCancelacion\Capsules\CancellationAnswer;
 use PhpCfdi\XmlCancelacion\Capsules\CapsuleInterface;
 use PhpCfdi\XmlCancelacion\Capsules\ObtainRelated;
-use PhpCfdi\XmlCancelacion\Definitions\CancelAnswer;
-use PhpCfdi\XmlCancelacion\Definitions\DocumentType;
-use PhpCfdi\XmlCancelacion\Definitions\RfcRole;
+use PhpCfdi\XmlCancelacion\Models\CancelAnswer;
+use PhpCfdi\XmlCancelacion\Models\DocumentType;
+use PhpCfdi\XmlCancelacion\Models\RfcRole;
 use PhpCfdi\XmlCancelacion\Exceptions\HelperDoesNotHaveCredentials;
 use PhpCfdi\XmlCancelacion\Signers\DOMSigner;
 use PhpCfdi\XmlCancelacion\Signers\SignerInterface;
