+++ tests/Unit/Capsules/CancellationTest.php
@@ -5,10 +5,10 @@ declare(strict_types=1);
 namespace PhpCfdi\XmlCancelacion\Tests\Unit\Capsules;
 
 use DateTimeImmutable;
-use PhpCfdi\XmlCancelacion\Capsules\CancelDocument;
-use PhpCfdi\XmlCancelacion\Capsules\CancelDocuments;
+use PhpCfdi\XmlCancelacion\Models\CancelDocument;
+use PhpCfdi\XmlCancelacion\Models\CancelDocuments;
 use PhpCfdi\XmlCancelacion\Capsules\Cancellation;
-use PhpCfdi\XmlCancelacion\Definitions\DocumentType;
+use PhpCfdi\XmlCancelacion\Models\DocumentType;
 use PhpCfdi\XmlCancelacion\Tests\TestCase;
 
 class CancellationTest extends TestCase
