+++ tests/Unit/Signers/XmlSecLibsSignerTest.php
@@ -7,8 +7,8 @@ namespace PhpCfdi\XmlCancelacion\Tests\Unit\Signers;
 use DateTimeImmutable;
 use Exception;
 use LogicException;
-use PhpCfdi\XmlCancelacion\Capsules\CancelDocument;
-use PhpCfdi\XmlCancelacion\Capsules\CancelDocuments;
+use PhpCfdi\XmlCancelacion\Models\CancelDocument;
+use PhpCfdi\XmlCancelacion\Models\CancelDocuments;
 use PhpCfdi\XmlCancelacion\Capsules\Cancellation;
 use PhpCfdi\XmlCancelacion\Credentials;
 use PhpCfdi\XmlCancelacion\Signers\DOMSigner;
