+++ tests/Unit/XmlCancelacionHelperSpy.php
@@ -6,10 +6,10 @@ namespace PhpCfdi\XmlCancelacion\Tests\Unit;
 
 use DateTimeImmutable;
 use LogicException;
-use PhpCfdi\XmlCancelacion\Capsules\CancelDocuments;
+use PhpCfdi\XmlCancelacion\Models\CancelDocuments;
 use PhpCfdi\XmlCancelacion\Capsules\Cancellation;
 use PhpCfdi\XmlCancelacion\Capsules\CapsuleInterface;
-use PhpCfdi\XmlCancelacion\Definitions\DocumentType;
+use PhpCfdi\XmlCancelacion\Models\DocumentType;
 use PhpCfdi\XmlCancelacion\XmlCancelacionHelper;
 
 class XmlCancelacionHelperSpy extends XmlCancelacionHelper
