+++ tests/Unit/XmlCancelacionHelperTest.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace PhpCfdi\XmlCancelacion\Tests\Unit;
 
 use DateTimeImmutable;
-use PhpCfdi\XmlCancelacion\Capsules\CancelDocument;
-use PhpCfdi\XmlCancelacion\Capsules\CancelDocuments;
+use PhpCfdi\XmlCancelacion\Models\CancelDocument;
+use PhpCfdi\XmlCancelacion\Models\CancelDocuments;
 use PhpCfdi\XmlCancelacion\Capsules\Cancellation;
 use PhpCfdi\XmlCancelacion\Capsules\CapsuleInterface;
 use PhpCfdi\XmlCancelacion\Credentials;
